Renamed getNonPersistedExpenseData to mergeClaimExpensesWithSubmittedResponses
